Simultaneous MouseWheel axis support and code cleanup - support both MouseWheel axis simultaneously - switch to preciseX and preciseY float for wheelmove and assume >=SDL-2.0.18 - switch to power of 2 GUI_Event enum constants - ilogxi clz based integer logarithm to reverse power of 2 constants from http://github.com/jopadan/ilogxi - clean up indent of enum constants - rename vec_t to scalar_t

+++ src/common/rendering/v_framebuffer.cpp
@@ -234,15 +234,15 @@ int DFrameBuffer::ScreenToWindowY(int y)
 	return mScreenViewport.top + mScreenViewport.height - (int)round(y * mScreenViewport.height / (float)GetHeight());
 }
 
-void DFrameBuffer::ScaleCoordsFromWindow(int16_t &x, int16_t &y)
+void DFrameBuffer::ScaleCoordsFromWindow(int32_t &x, int32_t &y)
 {
 	int letterboxX = mOutputLetterbox.left;
 	int letterboxY = mOutputLetterbox.top;
 	int letterboxWidth = mOutputLetterbox.width;
 	int letterboxHeight = mOutputLetterbox.height;
 
-	x = int16_t((x - letterboxX) * Width / letterboxWidth);
-	y = int16_t((y - letterboxY) * Height / letterboxHeight);
+	x = int32_t((x - letterboxX) * Width / letterboxWidth);
+	y = int32_t((y - letterboxY) * Height / letterboxHeight);
 }
 
 void DFrameBuffer::FPSLimit()
